Refactor addToCart logic
