+++ src/routes/(app)/shop/tickets/+page.server.ts
@@ -3,7 +3,7 @@ import { redirect } from "sveltekit-flash-message/server";
 import { message, superValidate } from "sveltekit-superforms/server";
 import { z } from "zod";
 import type { PageServerLoad } from "./$types";
-import { addTicketToCart, getTickets } from "./tickets";
+import { addTicketToCart, getTickets } from "./api/addToCart";
 
 export const load: PageServerLoad = async ({ locals }) => {
   const allTickets = await getTickets(locals.prisma);
