+++ src/routes/(app)/shop/tickets/api/mock.ts
